EID-941: Refactoring - moving classes to new packages
- CLI feels like separate logic from core logic processing so we've
created a separate package for that
- Moved KeyLoaders to utils because it's now used commonly between
trustanchor and metadata
