+++ build.gradle
@@ -26,7 +26,7 @@ apply plugin: 'application'
 apply plugin: 'maven-publish'
 apply plugin: 'org.junit.platform.gradle.plugin'
 
-mainClassName = 'uk.gov.ida.eidas.Application'
+mainClassName = 'uk.gov.ida.eidas.cli.Application'
 def buildVersion = "1.0-${System.env.BUILD_NUMBER ?: 'SNAPSHOT'}"
 version = "$buildVersion"
