+++ src/main/java/uk/gov/ida/eidas/cli/trustanchor/Print.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.trustanchor.cli;
+package uk.gov.ida.eidas.cli.trustanchor;
 
 import com.nimbusds.jose.JWSObject;
