+++ src/main/java/uk/gov/ida/eidas/cli/trustanchor/SignWithFile.java
@@ -1,18 +1,15 @@
-package uk.gov.ida.eidas.metadata.cli;
-
-import picocli.CommandLine.Command;
-import picocli.CommandLine.Option;
-import uk.gov.ida.eidas.trustanchor.FileKeyLoader;
+package uk.gov.ida.eidas.cli.trustanchor;
 
 import java.io.File;
 import java.security.cert.X509Certificate;
-import java.security.interfaces.ECPrivateKey;
 import java.security.interfaces.RSAPrivateKey;
 import java.util.concurrent.Callable;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
+import uk.gov.ida.eidas.utils.keyloader.FileKeyLoader;
 
 @Command(name="sign-with-file", description="Signs the final key set with a key loaded from a file")
 public class SignWithFile extends SigningCommand implements Callable<Void> {
-
   @Option(names = { "--key" }, description = "Location of the private key to use for signing", required=true)
   private File keyFile;
 
@@ -21,7 +18,7 @@ public class SignWithFile extends SigningCommand implements Callable<Void> {
 
   @Override
   public Void call() throws Exception {
-    ECPrivateKey key = FileKeyLoader.loadECKey(keyFile);
+    RSAPrivateKey key = FileKeyLoader.loadRSAKey(keyFile);
     X509Certificate x509Certificate = FileKeyLoader.loadCert(certificateFile);
     return build(key, x509Certificate);
   }
