+++ src/main/java/uk/gov/ida/eidas/cli/trustanchor/TrustAnchorGenerationApplication.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.trustanchor.cli;
+package uk.gov.ida.eidas.cli.trustanchor;
 
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
