+++ src/test/java/uk/gov/ida/eidas/metadata/ConnectorMetadataSignerTest.java
@@ -11,7 +11,7 @@ import org.opensaml.core.xml.io.UnmarshallingException;
 import org.opensaml.saml.common.SignableSAMLObject;
 import org.opensaml.xmlsec.signature.Signature;
 import org.opensaml.xmlsec.signature.support.SignatureConstants;
-import uk.gov.ida.eidas.trustanchor.FileKeyLoader;
+import uk.gov.ida.eidas.utils.keyloader.FileKeyLoader;
 import uk.gov.ida.eidas.utils.FileReader;
 
 import java.io.File;
